% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet-package.R
\docType{data}
\name{corona_data_all}
\alias{corona_data_all}
\title{Downloaded data for the vignette}
\format{
[data.frame] with all available stations in March 2020
}
\description{
Downloaded data for the vignette
}
\author{
Johannes Friedrich
}
