% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hystreet_request.R
\name{create_hystreet_request}
\alias{create_hystreet_request}
\title{Create HTTP request to the hystreet.com API}
\usage{
create_hystreet_request(endpoint, hystreetId = NULL, query = NULL, API_token)
}
\arguments{
\item{endpoint}{Character. Endpoint to send the request to.}

\item{hystreetId}{Integer. ID of a hystreet.com location.}

\item{query}{List. Query parameters to get data for specific date.}

\item{API_token}{Character. API key for accessing the hystreet.com API.}
}
\value{
A data.frame with parsed data from the hystreet.com API.
}
\description{
Create HTTP request to the hystreet.com API
}
