% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet-package.R
\docType{data}
\name{data_73_74}
\alias{data_73_74}
\title{Downloaded data for the vignette}
\format{
[data.frame] with pedestrians counta from January 2019
}
\description{
Downloaded data for the vignette
}
\author{
Johannes Friedrich
}
