% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convert_dates}
\alias{.convert_dates}
\title{Convert date into YMD-HMS format}
\usage{
.convert_dates(date)
}
\arguments{
\item{date}{Date to be converted}
}
\value{
Converted date
}
\description{
Convert date into YMD-HMS format
}
