% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hystreet_request.R
\name{.create_hystreet_request}
\alias{.create_hystreet_request}
\title{Create request to Hystreet API}
\usage{
.create_hystreet_request(API_key = NULL, hystreetId = NULL, query = NULL)
}
\arguments{
\item{API_key}{[character]: API key to get access to Hystreet API}

\item{hystreetId}{[integer]: Id of a hystreet location}

\item{query}{[list]: Querys to get data for specific date. Use with argument path = c("from", "to", "resolution").}
}
\value{
[data.frame] with parsed data from hystreet API
}
\description{
Create request to Hystreet API
}
\section{Function version}{

 0.0.2
}

\author{
Johannes Friedrich
}
\keyword{internal}
