% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hystreet_location_details.R
\name{get_hystreet_location_details}
\alias{get_hystreet_location_details}
\title{Retrieve details for a specific hystreet.com station}
\usage{
get_hystreet_location_details(
  hystreetId,
  API_token = Sys.getenv("HYSTREET_API_TOKEN")
)
}
\arguments{
\item{hystreetId}{Integer. Mandatory parameter. ID of the requested station. See `get_hystreet_locations()` for available station IDs.}

\item{API_token}{Character. API key for accessing the hystreet.com API.}
}
\value{
A list with detailed information on the station.
}
\description{
Detailed information in form of a named list includes, e.g., postal code, time of earliest and latest measurement and information on the sensor and the zones.
}
\examples{
\dontrun{
get_hystreet_location_details(148)
}

}
