% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hystreet_stats.R
\name{get_hystreet_stats}
\alias{get_hystreet_stats}
\title{Get statistics from the hystreet.com via hystreet.com's API}
\usage{
get_hystreet_stats(verbose = FALSE, API_token = NULL)
}
\arguments{
\item{verbose}{[logical] (**optional**): Show more verbose output? Defaults to FALSE.}

\item{API_token}{[character] (**optional**): API key to get access to hystreet.com API}
}
\value{
A [data.frame] with parsed data from hystreet.com API
}
\description{
Get statistics from the hystreet.com via hystreet.com's API
}
\examples{
\dontrun{
 get_hystreet_stats(TRUE)
}
}
