% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet-package.R
\docType{package}
\name{hystReet-package}
\alias{hystReet-package}
\title{hystReet}
\description{
An R API wrapper for the hystreet project. 
Download and analyse environmental data provided by https://hystreet.com/
}
\details{
\tabular{ll}{Package: \tab hystReet \cr Type: \tab Package \cr Version:
\tab 0.0.2 \cr Date: \tab 2020-06-14 \cr License: \tab GPL-2 \cr }
}
\keyword{package}
