% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet.R
\docType{package}
\name{hystReet}
\alias{hystReet-package}
\alias{hystReet}
\title{hystReet}
\description{
An R API wrapper for hystreet.com.  
Download and analyse data provided by <https://hystreet.com/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JohannesFriedrich/hystReet}
  \item Report bugs at \url{https://github.com/JohannesFriedrich/hystReet/issues}
}

}
\author{
\strong{Maintainer}: Johannes Friedrich \email{Johannes.Friedrich@posteo.de}

Other contributors:
\itemize{
  \item Buhl Yannik \email{yannik.buhl@posteo.de} [contributor]
}

}
\keyword{package}
