% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet-package.R
\docType{data}
\name{location_71}
\alias{location_71}
\title{Downloaded data for the vignette}
\format{
[list] with results from location 71 in december 2018
}
\description{
Downloaded data for the vignette
}
\author{
Johannes Friedrich
}
