% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{locations}
\alias{locations}
\title{Downloaded data for the vignette: All locations (March 2020)}
\format{
A data frame with location details.
}
\description{
A dataset with all available locations as of March 2020.
}
\keyword{datasets}
