% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_parameter_values}
\alias{parse_parameter_values}
\title{Parse query parameter list}
\usage{
parse_parameter_values(query)
}
\arguments{
\item{query}{Named query parameter list}
}
\value{
Parsed query list
}
\description{
Parse query parameter list
}
