% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet-package.R
\docType{data}
\name{ratio}
\alias{ratio}
\title{Downloaded data for the vignette}
\format{
[data.frame] with pedestrians counta/day
}
\description{
Downloaded data for the vignette
}
\author{
Johannes Friedrich
}
