% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hystreet_token.R
\name{set_hystreet_token}
\alias{set_hystreet_token}
\title{Set your API token for the hystreet.com API}
\usage{
set_hystreet_token(API_token)
}
\arguments{
\item{API_token}{Character. API key to get access to hystreet.com API.}
}
\value{
Environment variable set to the API token
}
\description{
Set your API token for the hystreet.com API
}
