% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_api_token}
\alias{valid_api_token}
\title{Check validity of API token}
\usage{
valid_api_token(API_token)
}
\arguments{
\item{API_token}{Token to be checked}
}
\description{
Check validity of API token
}
