% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{corona_data_all}
\alias{corona_data_all}
\title{Downloaded data for the vignette: All stations (March 2020)}
\format{
A data frame with station data.
}
\description{
A dataset with data from all available stations in March 2020.
}
\keyword{datasets}
