% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data_73_74}
\alias{data_73_74}
\title{Data for the vignette: Pedestrian counts (Jan 2019)}
\format{
A data frame with pedestrian counts.
}
\description{
A dataset containing pedestrian counts from January 2019.
}
\keyword{datasets}
