% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hystreet_locations.R
\name{get_hystreet_locations}
\alias{get_hystreet_locations}
\title{Get station IDs of the Hystreet project}
\usage{
get_hystreet_locations(API_token = NULL)
}
\arguments{
\item{API_token}{\link{character} (\strong{optional}): API key to get access to Hystreet API}
}
\value{
\link{data.frame} with parsed data from hystreet API
}
\description{
Get station IDs of the Hystreet project
}
\section{Function version 0.0.1}{
NA
}

\examples{
\dontrun{
 all_hystreet_IDs <- get_hystreet_locations()
}
}
\author{
Johannes Friedrich
}
