% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{location_71}
\alias{location_71}
\title{Downloaded data for the vignette: Location 71 (Dec 2018)}
\format{
A list with location data.
}
\description{
A list containing results from location 71 in December 2018.
}
\keyword{datasets}
