% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hystReet-package.R
\docType{data}
\name{locations}
\alias{locations}
\title{Downloaded data for the vignette}
\format{
[data.frame] with all available locaations (03/27/2020)
}
\description{
Downloaded data for the vignette
}
\author{
Johannes Friedrich
}
