% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ratio}
\alias{ratio}
\title{Downloaded data for the vignette: Pedestrian counts per day}
\format{
A data frame with pedestrian counts per day.
}
\description{
A dataset with daily pedestrian counts.
}
\keyword{datasets}
