% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hystreet_csv.R
\name{read_hystreet_csv}
\alias{read_hystreet_csv}
\title{Read downloaded data from the hystreet.com web page to data.frames}
\usage{
read_hystreet_csv(path, ...)
}
\arguments{
\item{path}{[character] (**required**): Path to the downloaded CSV file}

\item{...}{(**optional**): Arguments passed to [utils::read.csv2()]}
}
\value{
A [data.frame] with parsed data from hystreet.com API
}
\description{
Read downloaded data from the hystreet.com web page to data.frames
}
\examples{
\dontrun{
df <- hystReet::read_hystreet_csv("~/Downloads/hystreet.csv")
 }
}
