% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hystreet_csv.R
\name{read_hystreet_csv}
\alias{read_hystreet_csv}
\title{Read downloaded data from the hystreet webpage to  data.frames}
\usage{
read_hystreet_csv(path, ...)
}
\arguments{
\item{path}{\link{string} (\strong{required}): path to the downloaded csv file}

\item{...}{(\strong{optional}): arguments passed to \code{\link[utils:read.table]{utils::read.csv2()}}}
}
\value{
\link{data.frame} with parsed data from hystreet API
}
\description{
Read downloaded data from the hystreet webpage to  data.frames
}
\section{Function version}{

0.0.1
}

\examples{
\dontrun{
## read data from downlaoded csv file data
df <- hystReet::read_hystreet_csv("~/Downloads/hystreet.csv")

 }
}
\author{
Johannes Friedrich
}
