% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hystreet_token.R
\name{set_hystreet_token}
\alias{set_hystreet_token}
\title{Set your API token for the Hystreet API}
\usage{
set_hystreet_token(API_token = NULL)
}
\arguments{
\item{API_token}{\link{character} (\strong{required}): API key to get access to Hystreet API}
}
\value{
Set environment variable
}
\description{
Set your API token for the Hystreet API
}
\section{Function version}{

0.0.1
}

\author{
Johannes Friedrich
}
